Type part of DNA Accessibility state and settings component (#31)
* Move some dna accessibility state types to typed file

* Update imports and type dna accessibility state

* lint
