+++ src/utils/string.ts
@@ -1,0 +1,3 @@
+export function capitalizeFirstLetter(s: string) {
+  return s.charAt(0).toUpperCase() + s.slice(1).toLowerCase();
+}
